Create  FizzBuzz
Difficulty: Easy

Problem Statement:
Write a function that takes an integer n as input. The function should print numbers from 1 to n. However, there are special rules:

For multiples of three, print "Fizz" instead of the number.

For multiples of five, print "Buzz" instead of the number.

For numbers which are multiples of both three and five, print "FizzBuzz" instead of the number.

For all other numbers, print the number itself.

Examples:

If n = 5, the output should be:

1
2
Fizz
4
Buzz
If n = 15, the output should be:

1
2
Fizz
4
Buzz
Fizz
7
8
Fizz
Buzz
11
Fizz
13
14
FizzBuzz
Constraints:

n will be a positive integer.

Tips for Solving:

Iteration: How will you generate numbers from 1 to n?

Order of Checks: This is crucial! If a number is a multiple of both three and five, it's also a multiple of three, and it's also a multiple of five. What order should your if/elif/else statements be in to ensure "FizzBuzz" is printed for numbers like 15?

Modulus Operator: Remember the modulus operator (%) to check for divisibility. number % divisor == 0 means number is a multiple of divisor.